Add ec2 instances list
