+++ CloudyCider/views/components/EC2/EC2.swift
